+++ app/components/UI/PermissionsSummary/PermissionsSummary.test.tsx
@@ -3,48 +3,13 @@ import PermissionsSummary from './PermissionsSummary';
 import { backgroundState } from '../../../util/test/initial-root-state';
 import renderWithProvider from '../../../util/test/renderWithProvider';
 import { MOCK_ACCOUNTS_CONTROLLER_STATE } from '../../../util/test/accountsControllerTestUtils';
+import { KeyringTypes } from '@metamask/keyring-controller';
+import {
+  AvatarSize,
+  AvatarVariant,
+} from '../../../component-library/components/Avatars/Avatar/Avatar.types';
 
 const mockedNavigate = jest.fn();
-
-// Mock useSelectedAccount hook
-jest.mock('../Tabs/TabThumbnail/useSelectedAccount', () => ({
-  __esModule: true,
-  default: jest.fn().mockReturnValue({
-    name: 'Account 2',
-    address: '0x0',
-    isSelected: true,
-    assets: {
-      fiatBalance: '$3200',
-    },
-  }),
-}));
-
-// Mock useAccounts hook
-jest.mock('../../../components/hooks/useAccounts', () => ({
-  useAccounts: jest.fn().mockReturnValue({
-    accounts: [
-      {
-        name: 'Account 2',
-        address: '0x0',
-        isSelected: true,
-        assets: {
-          fiatBalance: '$3200',
-        },
-      },
-    ],
-    evmAccounts: [
-      {
-        name: 'Account 2',
-        address: '0x0',
-        isSelected: true,
-        assets: {
-          fiatBalance: '$3200',
-        },
-      },
-    ],
-    ensByAccountAddress: {},
-  }),
-}));
 
 jest.mock('@react-navigation/native', () => {
   const actualNav = jest.requireActual('@react-navigation/native');
@@ -55,6 +20,14 @@ jest.mock('@react-navigation/native', () => {
     }),
   };
 });
+
+jest.mock('react-native-scrollable-tab-view', () => ({
+  __esModule: true,
+  default: ({ children }: { children: React.ReactNode }) => <>{children}</>,
+  DefaultTabBar: ({ children }: { children: React.ReactNode }) => (
+    <>{children}</>
+  ),
+}));
 
 const mockInitialState = {
   wizard: {
@@ -82,6 +55,7 @@ describe('PermissionsSummary', () => {
           chainId: '0x1',
         }}
         isNetworkSwitch
+        accounts={[]}
       />,
       { state: mockInitialState },
     );
@@ -95,6 +69,29 @@ describe('PermissionsSummary', () => {
           icon: '',
           url: 'https://app.uniswap.org/',
         }}
+        accounts={[
+          {
+            name: 'Account 2',
+            address: '0x2',
+            isSelected: true,
+            assets: {
+              fiatBalance: '$3200',
+            },
+            caipAccountId: 'eip155:0:0x2',
+            yOffset: 0,
+            type: KeyringTypes.simple,
+          },
+        ]}
+        accountAddresses={['eip155:0:0x2']}
+        networkAvatars={[
+          {
+            name: 'Ethereum Mainnet',
+            // eslint-disable-next-line @typescript-eslint/no-require-imports
+            imageSource: require('../../../assets/images/network-avatar.png'),
+            size: AvatarSize.Xs,
+            variant: AvatarVariant.Network,
+          },
+        ]}
       />,
       { state: mockInitialState },
     );