+++ app/components/Views/confirmations/components/info-root/info-root.tsx
@@ -4,6 +4,7 @@ import React from 'react';
 import { UnstakeConfirmationViewProps } from '../../../../UI/Stake/Views/UnstakeConfirmationView/UnstakeConfirmationView.types';
 import { useQRHardwareContext } from '../../context/qr-hardware-context';
 import useApprovalRequest from '../../hooks/useApprovalRequest';
+import { use7702TransactionType } from '../../hooks/7702/use7702TransactionType';
 import { useTransactionMetadataRequest } from '../../hooks/transactions/useTransactionMetadataRequest';
 import ContractInteraction from '../info/contract-interaction';
 import PersonalSign from '../info/personal-sign';
@@ -11,6 +12,7 @@ import QRInfo from '../qr-info';
 import StakingClaim from '../../external/staking/info/staking-claim';
 import StakingDeposit from '../../external/staking/info/staking-deposit';
 import StakingWithdrawal from '../../external/staking/info/staking-withdrawal';
+import SwitchAccountType from '../info/switch-account-type';
 import Transfer from '../info/transfer';
 import TypedSignV1 from '../info/typed-sign-v1';
 import TypedSignV3V4 from '../info/typed-sign-v3v4';
@@ -32,6 +34,8 @@ const ConfirmationInfoComponentMap = {
     transactionType,
   }: ConfirmationInfoComponentRequest) => {
     switch (transactionType) {
+      case TransactionType.batch:
+        return ContractInteraction;
       case TransactionType.contractInteraction:
         return ContractInteraction;
       case TransactionType.stakingClaim:
@@ -58,9 +62,14 @@ const Info = ({ route }: InfoProps) => {
   const { approvalRequest } = useApprovalRequest();
   const transactionMetadata = useTransactionMetadataRequest();
   const { isSigningQRObject } = useQRHardwareContext();
+  const { isDowngrade, isUpgradeOnly } = use7702TransactionType();
 
   if (!approvalRequest?.type) {
     return null;
+  }
+
+  if (isDowngrade || isUpgradeOnly) {
+    return <SwitchAccountType />;
   }
 
   if (isSigningQRObject) {