+++ app/util/networks/index.js
@@ -9,6 +9,7 @@ import {
   LINEA_MAINNET,
   LINEA_SEPOLIA,
   MEGAETH_TESTNET,
+  MONAD_TESTNET,
 } from '../../../app/constants/network';
 import { NetworkSwitchErrorType } from '../../../app/constants/error';
 import {
@@ -28,6 +29,7 @@ const sepoliaLogo = require('../../images/sepolia-logo-dark.png');
 const lineaTestnetLogo = require('../../images/linea-testnet-logo.png');
 const lineaMainnetLogo = require('../../images/linea-mainnet-logo.png');
 const megaEthTestnetLogo = require('../../images/megaeth-testnet-logo.png');
+const monadTestnetLogo = require('../../images/monad-testnet-logo.png');
 
 /* eslint-enable */
 import {
@@ -59,7 +61,7 @@ import {
 } from '../../selectors/multichainNetworkController';
 import { formatBlockExplorerAddressUrl } from '../../core/Multichain/networks';
 import { CHAIN_IDS } from '@metamask/transaction-controller';
-import { isCaipChainId } from '@metamask/utils';
+import { isCaipChainId, KnownCaipNamespace, parseCaipChainId } from '@metamask/utils';
 
 /**
  * List of the supported networks
@@ -133,6 +135,19 @@ export const NetworkList = {
     networkType: 'megaeth-testnet',
     imageSource: megaEthTestnetLogo,
     blockExplorerUrl: BlockExplorerUrl['megaeth-testnet'],
+  },
+  [MONAD_TESTNET]: {
+    name: 'Monad Testnet',
+    shortName: 'Monad Testnet',
+    networkId: 10143,
+    chainId: toHex('10143'),
+    ticker: 'MON',
+    // Third party color
+    // eslint-disable-next-line @metamask/design-tokens/color-no-hex
+    color: '#61dfff',
+    networkType: 'monad-testnet',
+    imageSource: monadTestnetLogo,
+    blockExplorerUrl: BlockExplorerUrl['monad-testnet'],
   },
   [RPC]: {
     name: 'Private Network',
@@ -246,6 +261,9 @@ export const getTestNetImageByChainId = (chainId) => {
   if (NETWORKS_CHAIN_ID.MEGAETH_TESTNET === chainId) {
     return networksWithImages?.['MEGAETH-TESTNET'];
   }
+  if (NETWORKS_CHAIN_ID.MONAD_TESTNET === chainId) {
+    return networksWithImages?.['MONAD-TESTNET'];
+  }
 };
 
 /**
@@ -257,6 +275,7 @@ export const TESTNET_CHAIN_IDS = [
   ChainId[NetworkType['linea-goerli']],
   ChainId[NetworkType['linea-sepolia']],
   ChainId[NetworkType['megaeth-testnet']],
+  ChainId[NetworkType['monad-testnet']],
 ];
 
 /**
@@ -490,14 +509,14 @@ export const getNetworkNameFromProviderConfig = (providerConfig) => {
 };
 
 /**
- * Gets the image source given both the network type and the chain ID.
+ * Gets the image source for an EVM network given both the network type and the Hex chain ID.
  *
  * @param {object} params - Params that contains information about the network.
  * @param {string=} params.networkType - Type of network from the provider.
- * @param {string} params.chainId - ChainID of the network.
+ * @param {string} params.chainId - Hex EVM chain ID of the EVM network.
  * @returns {Object} - Image source of the network.
  */
-export const getNetworkImageSource = ({ networkType, chainId }) => {
+const getEvmNetworkImageSource = ({ networkType, chainId }) => {
   const defaultNetwork = getDefaultNetworkByChainId(chainId);
 
   if (defaultNetwork) {
@@ -522,11 +541,28 @@ export const getNetworkImageSource = ({ networkType, chainId }) => {
     return customNetworkImg;
   }
 
+  return getTestNetImage(networkType);
+};
+
+/**
+ * Gets the image source for a network given both the network type and the Hex EVM chain ID or CaipChainId.
+ *
+ * @param {object} params - Params that contains information about the network.
+ * @param {string=} params.networkType - Type of network from the provider.
+ * @param {string} params.chainId - Hex EVM chain ID or CAIP chain ID of the network.
+ * @returns {Object} - Image source of the network.
+ */
+export const getNetworkImageSource = ({ networkType, chainId }) => {
+  let hexChainId = chainId;
   if (isCaipChainId(chainId)) {
-    return getNonEvmNetworkImageSourceByChainId(chainId);
+    const {namespace, reference} = parseCaipChainId(chainId);
+    if (namespace !== KnownCaipNamespace.Eip155) {
+      return getNonEvmNetworkImageSourceByChainId(chainId);
+    }
+    hexChainId = toHex(reference);
   }
 
-  return getTestNetImage(networkType);
+  return getEvmNetworkImageSource({ networkType, chainId: hexChainId});
 };
 
 /**
@@ -597,6 +633,9 @@ export const isChainPermissionsFeatureEnabled = true;
 export const isPermissionsSettingsV1Enabled =
   process.env.MM_PERMISSIONS_SETTINGS_V1_ENABLED === 'true';
 
+export const isPerDappSelectedNetworkEnabled = () =>
+  process.env.MM_PER_DAPP_SELECTED_NETWORK === 'true';
+
 export const isPortfolioViewEnabled = () =>
   process.env.PORTFOLIO_VIEW === 'true';
 
@@ -607,6 +646,7 @@ export const WHILELIST_NETWORK_NAME = {
   [ChainId.mainnet]: 'Mainnet',
   [ChainId['linea-mainnet']]: 'Linea Mainnet',
   [ChainId['megaeth-testnet']]: 'Mega Testnet',
+  [ChainId['monad-testnet']]: 'Monad Testnet',
 };
 
 /**